Major Internal Refactor of ObservableList:
ObservableList is now a generic version of ObservableIList<T,List<T>> instead of its own implementation.

These two classes used to have significant differences (e.g. AddRange was not initially in ObservableIList) the code base converged and they are now almost the same. They were benchmarked at the same speed.

The only instrinsic difference is that AddRange(...) is now calls a virtual method AddRangeInternal(...) that is overridden in the List<T> version. The extra jump to an override is neglible compared to the time of the Event calls.

Changes To ObservableIList:
The internal `_list` method has also been set to protected instead of private.

AddRangeInternal, AddMoveInternal are now used in AddRange(...) and AddMove(...). They call an internal overridable implementation of these operations.

Nuance:
fixed spacing/order/comments in IObservableList
